refactor(metrics): extract LowestFee's resize walk into a documented helper
Move the "resize trick" out of `LowestFee::bound` into a private
`resize_bound` helper in `metrics.rs`, returning a `ResizeBound` enum so the
caller keeps its own exact-match computation. No behavior change.

The main point is the helper's doc: the walk's crossing point and `scale`
embed the greedy prefix's actual (`input_weight()`-corrected, vbyte-rounded)
fee and weight, which a correction-avoiding descendant can undercut — so it
carries an explicit CAUTION that it must not be used as an `input_weight`
lower bound (`ChangelessWaste::bound` uses raw-weight-linear arithmetic for
exactly that reason). The same caveat applies in fee space to the remaining
use here; that pre-existing `LowestFee` issue is tracked upstream in #56.
